requirements.txt
# Camaral RAG Chatbot — dependencies
# Python 3.11+ recommended

# Agent & LLM
pydantic-ai-slim[google]>=0.0.14
google-genai>=1.0.0

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
pgvector>=0.3.0

# Config
pydantic-settings>=2.0.0

# Telegram
python-telegram-bot>=21.0

# Observability
logfire>=0.40.0

# PDF ingestion
pymupdf>=1.24.0

# HTTP (for google-genai / optional)
httpx>=0.27.0

# Tests
pytest>=8.0.0
pytest-asyncio>=0.24.0
